Refactor, refine, docs
